Implementations for 1.1.0: Separated constructors for Client & Server (To later implement client-side) Removed SQL dependency over chunk claiming (Now saved to chunk NBT) Removed SQL dependency over player claim options (Now saves to custom NBT) Removed SQL dependency over towns (Now saves to custom NBT) Removed SQL dependency for money (Saves to player NBT.dat) Fixed bed spawns not moving in "/setspawn" command Fixed iron doors from being opened if placed next to a wooden door Fixed death chests only spawning on the surface, now checking for CAVE_AIR Added command to convert from legacy database storage Added saving of custom NBT files to built-in server saving Added chunk position handling for all 16x16x264 blocks in a chunk, for shops Added new wilderness names for nether & end Added a new class for giving and taking money for players, online or offline Added daily reward-money for logging in (Configurable) Added chat message for death chest coordinates (Configurable) Added a version to the config file for porting Added translation keys for money Added separate global/town/local chat Added command for changing chats Added new fun option that passive mobs will seek shelter from rain Modified WitherBoss sound to only be localized

+++ src/main/java/net/TheElm/project/mixins/Player/Interaction/ShopSign.java
@@ -26,7 +26,6 @@
 package net.TheElm.project.mixins.Player.Interaction;
 
 import net.TheElm.project.CoreMod;
-import net.TheElm.project.config.SewingMachineConfig;
 import net.TheElm.project.interfaces.ShopSignBlockEntity;
 import net.TheElm.project.utilities.ShopSignBuilder;
 import net.minecraft.block.entity.BlockEntity;
@@ -64,7 +63,7 @@ public abstract class ShopSign extends BlockEntity implements ShopSignBlockEntit
     private UUID shopSign_Owner = null;
     private Identifier shopSign_item = null;
     private Integer shopSign_itemCount = null;
-    private Long shopSign_itemPrice = null;
+    private Integer shopSign_itemPrice = null;
     
     /*
      * Mixin Getters
@@ -90,7 +89,7 @@ public abstract class ShopSign extends BlockEntity implements ShopSignBlockEntit
         return this.shopSign_itemCount;
     }
     @Override @Nullable
-    public Long getShopItemPrice() {
+    public Integer getShopItemPrice() {
         return this.shopSign_itemPrice;
     }
     @Override
@@ -150,8 +149,8 @@ public abstract class ShopSign extends BlockEntity implements ShopSignBlockEntit
             tag.putString("shop_item_mod", this.shopSign_item.getNamespace());
             tag.putString("shop_item_name", this.shopSign_item.getPath());
         }
-        tag.putLong( "shop_item_count", this.shopSign_itemCount );
-        tag.putLong( "shop_price", this.shopSign_itemPrice );
+        tag.putInt( "shop_item_count", this.shopSign_itemCount );
+        tag.putInt( "shop_price", this.shopSign_itemPrice );
         
         callback.setReturnValue( tag );
     }
@@ -164,7 +163,7 @@ public abstract class ShopSign extends BlockEntity implements ShopSignBlockEntit
                 this.shopSign_item = new Identifier( signItem );
                 this.shopSign_Owner = tag.getUuid("shop_owner");
                 this.shopSign_itemCount = tag.getInt("shop_item_count");
-                this.shopSign_itemPrice = tag.getLong("shop_price");
+                this.shopSign_itemPrice = tag.getInt("shop_price");
             } catch (InvalidIdentifierException e) {
                 CoreMod.logError( "Could not find item \"" + signItem + "\" for shop sign." );
                 CoreMod.logError( e );
